Authorization with devise gem
This feature adds basic authorization to the app with the Devise gem.

Authentication is needed for this app so that admins (ie. DPLA staff)
can log in to perform CRUD actions in the database. After our initial
launch in October, non-logged-in users will be able to view sets,
sources, and guides; but until launch, only admins should have the
ability to access any part of the application. So for now, I added
before_action :authenticate_admin! to the ApplicationController. When we
launch in October, this will need to change.

I kept most of the default Devise settings, expressed in the database
migration, config/initializers/devise.rb and
config/locales/devise.en.yml. One exception is that I added the DPLA
email address in line 13 of config/initializers/devise.rb.

The setting for the Admin model are very simple. I am double-checking
with Franky, but I think we only need one admin account that DPLA staff
members can share, and its easy to make a new account from the command
line. So there is log-in and log-out capability from the app interfaces,
but there is NOT ability to make a new account or reset a password. We
can always add these things in later if we decide we need them.

I made some spec settings changes to work with Devise: see
spec/support/devise.rb and spec/support/controller_macros.rb and the new
factory, spec/factories/admins.rb. Although there appears to be a lot of
changes to the existing specs, it's really just nesting all of the
existing tests within a context block that logs in the admin.

I was also getting some errors with the rspec that I fixed by adding
.rspec and spec/rails_helper.rb. These files are auto-generated with you
do rails generate rspec:install, which up to now I had neglected to do.