Add media uploads
Add s3_browser_uploads gem and code to handle uploads.  I've added a
class of our own, PSSBrowserUploads::FormDefinition, which extends its
base class in S3BrowserUploads, in order to get it to work with the
us-east-1 region.

Remove mime_type field from Video and Audio models, because we won't
need to store it.

I don't know how to make tests for the Javascript code.  That will be
done eventually.

Additionally:

* Break out tests for redirecting and re-rendering the #new view
  to separate shared examples for inclusion by only the controllers
  that need them.
* Create a new slot in the <head> of the application layout for inline
  <script>s
* Remove turbolinks, which breaks events