Add basic account management
* Add RegistrationsController, which overrides
  Devise::RegistrationsController

* Add account-confirmation features to Admin model, including schema
  changes

* Update Vagrantfile and README.md for use of mailcatcher in development

* Add Devise views

* Change the registration (account) #update to not check account's
  existing password.  This allows a manager to edit another user's
  account without knowing the account's password.

* Remove "current password" field from update view

* Fix styles of login-related pages

Fix existing tests to work with registrations:
* Add factory for invalid Admin creation fields
* Adjust "#index sets resources variable" shared example in "basic
  controller" to work with there being multiple rows in Admin after the
  manager is logged in and the `create' macro creates another admin.

Instead of creating our own controller for account management, reuse
Devise::RegistrationsController, overriding what we need to customise.

Override the Devise behavior to allow administrators to edit other
users' accounts.  Also, change the default behavior so that only admins
can create accounts.

The user has to be authenticated to view registration (account
management) pages.

Add confirmed_at and confirmation_sent_at values to Admin factory to
keep Devise from redirecting actions to the confirmation controller,
when it thinks the account has not been confirmed.  This was causing
tests to fail.

The destinations of redirects have been changed in various places from
the Devise defaults.  For example, if an account is created, the
redirect now goes back to the index of registrations (accounts) instead
of the home page.